+++ lib/plausible_web/router.ex
@@ -89,6 +89,7 @@ defmodule PlausibleWeb.Router do
     delete "/:site_id", ExternalSitesController, :delete_site
     put "/shared-links", ExternalSitesController, :find_or_create_shared_link
     put "/goals", ExternalSitesController, :find_or_create_goal
+    put "/goals/:goal_id", ExternalSitesController, :update_goal
     delete "/goals/:goal_id", ExternalSitesController, :delete_goal
   end