+++ test/plausible_web/controllers/api/external_sites_controller_test.exs
@@ -308,6 +308,71 @@ defmodule PlausibleWeb.Api.ExternalSitesControllerTest do
     end
   end
 
+  describe "PUT /api/v1/sites/goals/:goal_id" do
+    setup :create_new_site
+
+    test "update a goal by it's id", %{conn: conn, site: site} do
+      conn =
+        put(conn, "/api/v1/sites/goals", %{
+          site_id: site.domain,
+          goal_type: "event",
+          goal_value: "Signup"
+        })
+
+      %{"goal_id" => goal_id} = json_response(conn, 200)
+
+      conn = put(conn, "/api/v1/sites/goals/#{goal_id}", %{
+        site_id: site.domain,
+        goal_type: "event",
+        goal_value: "Signin"
+      })
+
+      assert json_response(conn, 200) == %{"updated" => true}
+    end
+
+    test "is 404 when goal cannot be found", %{conn: conn, site: site} do
+      conn = put(conn, "/api/v1/sites/goals/0", %{
+        site_id: site.domain,
+        goal_type: "event",
+        goal_value: "Signin"
+      })
+
+      assert json_response(conn, 404) == %{"error" => "Site or Goal could not be found"}
+    end
+
+    test "cannot update a goal belongs to a site that the user does not own", %{conn: conn, user: user} do
+      site = insert(:site, members: [])
+      insert(:site_membership, user: user, site: site, role: :admin)
+      
+      conn = put(conn, "/api/v1/sites/goals/1", %{
+        site_id: site.domain,
+        goal_type: "event",
+        goal_value: "Signin"
+      })
+
+      assert json_response(conn, 404) == %{"error" => "Site or Goal could not be found"}
+    end
+
+    test "cannot access with a bad API key scope", %{conn: conn, site: site, user: user} do
+      api_key = insert(:api_key, user: user, scopes: ["stats:read:*"])
+
+      conn =
+        conn
+        |> Plug.Conn.put_req_header("authorization", "Bearer #{api_key.key}")
+      
+      conn = put(conn, "/api/v1/sites/goals/1", %{
+        site_id: site.domain,
+        goal_type: "event",
+        goal_value: "Signin"
+      })
+
+      assert json_response(conn, 401) == %{
+               "error" =>
+                 "Invalid API key. Please make sure you're using a valid API key with access to the resource you've requested."
+             }
+    end
+  end
+  
   describe "DELETE /api/v1/sites/goals/:goal_id" do
     setup :create_new_site
 
@@ -321,7 +386,7 @@ defmodule PlausibleWeb.Api.ExternalSitesControllerTest do
 
       %{"goal_id" => goal_id} = json_response(conn, 200)
 
-      conn = delete(conn, "/api/v1/sites/goals/" <> goal_id, %{
+      conn = delete(conn, "/api/v1/sites/goals/#{goal_id}", %{
         site_id: site.domain
       })
 
@@ -339,16 +404,8 @@ defmodule PlausibleWeb.Api.ExternalSitesControllerTest do
     test "cannot delete a goal belongs to a site that the user does not own", %{conn: conn, user: user} do
       site = insert(:site, members: [])
       insert(:site_membership, user: user, site: site, role: :admin)
-      conn =
-        put(conn, "/api/v1/sites/goals", %{
-          site_id: site.domain,
-          goal_type: "event",
-          goal_value: "Signup"
-        })
 
-      %{"goal_id" => goal_id} = json_response(conn, 200)
-
-      conn = delete(conn, "/api/v1/sites/goals/" <> goal_id, %{
+      conn = delete(conn, "/api/v1/sites/goals/1", %{
         site_id: site.domain
       })